DEV: Mostly just Ribbon testing
Got the custom button to show up on the Attachments tab of the
Message.Read Inspector, though! Had to muck in the XML -- the
Designer apparently doesn't play well with Tool Tabs.

Still need to create the Ribbon stuff for reattaching.

Need to turn on Option Strict and clean up the huge mess of
violations there.